Key vault signing
Looks like it's not fundamentally possible to do this, because an app cannot get
tokens for interactig with the key vault. User interaction is still possible but
the basic concept appears to be irrelevant anyway, since a user's access to
their own certificate is best guarded by their own credentials. This approach is
not TNO.